Restructure tests to make them visible by the test coverage tool
