+++ pkg/otel/metrics/otlp_to_arrow_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package metrics_test
+package metrics
 
 import (
 	"testing"
@@ -21,7 +21,6 @@ import (
 	"github.com/f5/otel-arrow-adapter/pkg/air/config"
 	"github.com/f5/otel-arrow-adapter/pkg/arrow"
 	"github.com/f5/otel-arrow-adapter/pkg/datagen"
-	"github.com/f5/otel-arrow-adapter/pkg/otel/metrics"
 )
 
 func TestOtlpMetricsToArrowRecords(t *testing.T) {
@@ -31,14 +30,14 @@ func TestOtlpMetricsToArrowRecords(t *testing.T) {
 	rr := air.NewRecordRepository(cfg)
 	lg := datagen.NewMetricsGenerator(datagen.DefaultResourceAttributes(), datagen.DefaultInstrumentationScopes())
 
-	multivariateConf := metrics.MultivariateMetricsConfig{
+	multivariateConf := MultivariateMetricsConfig{
 		Metrics: make(map[string]string),
 	}
 	multivariateConf.Metrics["system.cpu.time"] = "state"
 	multivariateConf.Metrics["system.memory.usage"] = "state"
 
 	request := lg.Generate(10, 100)
-	multiSchemaRecords, err := metrics.OtlpMetricsToArrowRecords(rr, request, &multivariateConf, cfg)
+	multiSchemaRecords, err := OtlpMetricsToArrowRecords(rr, request, &multivariateConf, cfg)
 	if err != nil {
 		t.Errorf("Unexpected error: %v", err)
 	}
